Refactor: Component files structure
