+++ client/src/components/Course/index.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import CourseService from '../services/course.service';
+import CourseService from '../../services/course.service';
 
 const CourseComponent = (props) => {
   let { currentUser, setCurrentUser } = props;
