+++ client/src/components/Login/index.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useHistory } from 'react-router';
-import AuthService from '../services/auth.service';
+import AuthService from '../../services/auth.service';
 
 const LoginComponent = (props) => {
   let { currentUser, setCurrentUser } = props;
