+++ client/src/components/Nav/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link, useHistory } from 'react-router-dom';
-import AuthService from '../services/auth.service';
+import AuthService from '../../services/auth.service';
 
 const NavComponent = (props) => {
   let { currentUser, setCurrentUser } = props;
