+++ client/src/components/PostCourse/index.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import CourseService from '../services/course.service';
+import CourseService from '../../services/course.service';
 
 const PostCourseComponent = (props) => {
   let { currentUser, setCurrentUser } = props;
@@ -39,7 +39,10 @@ const PostCourseComponent = (props) => {
       {!currentUser && (
         <div>
           <p>You must login first before seeing posts.</p>
-          <button className='btn btn-primary btn-lg' onClick={handleTakeToLogin}>
+          <button
+            className='btn btn-primary btn-lg'
+            onClick={handleTakeToLogin}
+          >
             Take me to login page.
           </button>
         </div>
