+++ client/src/components/Profile/index.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import AuthService from '../services/auth.service';
+import AuthService from '../../services/auth.service';
 
 const ProfileComponent = (props) => {
   let { currentUser, setCurrentUser } = props;
